+++ workspaces/ai-integrations/plugins/ai-experience/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @red-hat-developer-hub/backstage-plugin-ai-experience
 
+## 0.4.0
+
+### Minor Changes
+
+- 44c967a: Backstage version bump to v1.44.2
+
+### Patch Changes
+
+- Updated dependencies [44c967a]
+  - @red-hat-developer-hub/backstage-plugin-ai-experience-backend@0.3.0
+  - @red-hat-developer-hub/backstage-plugin-ai-experience-common@0.3.0
+
 ## 0.3.1
 
 ### Patch Changes